requirements.txt
azure-functions
haikunator
applicationinsights
azure-cli-core
azure-mgmt-resource
